+++ Bank.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace linq
+{
+
+    //Define a class for banks
+    public class Bank
+    {
+        public string Name { get; set; }
+        public int MillionaireCount { get; set; }
+        public string Symbol { get; set; }
+    }
+}
