+++ Customer.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace linq
+{
+    public class Customer
+    {
+        public string Name { get; set; }
+        public double Balance { get; set; }
+        public string Bank { get; set; }
+    }
+}
